Throw an exception when response is not a string (issue #51).
