+++ tests/fXmlRpc/Parser/XmlReaderParserTest.php
@@ -225,4 +225,17 @@ class XmlReaderParserTest extends AbstractParserTest
         );
         $this->parser->parse($string, $isFault);
     }
+
+    public function testThrowExceptionWhenIsString()
+    {
+        $string = 'returned string';
+
+        $isFault = true;
+
+        $this->setExpectedException(
+            'fXmlRpc\Exception\ParserException',
+            'Invalid XML. Expected XML, string given: "returned string"'
+        );
+        $this->parser->parse($string, $isFault);
+    }
 }
